LPS-158489 Inline LiferayMimeMessage into MailEngine as inner class
